pyproject.toml
[project]
name = "authapp-demo"
version = "0.1.0"
description = "Authentication & Authorization Demo Application - Simulate SAML, OAuth 2.0, and OpenID Connect flows"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.0.0",
    "flask-session>=0.5.0",
    "msal>=1.24.0",
    "requests>=2.31.0",
    "python-jose[cryptography]>=3.3.0",
    "cryptography>=41.0.0",
    "python-dotenv>=1.0.0",
    "pyjwt>=2.8.0",
    "gunicorn>=21.2.0"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project.urls]
Homepage = "https://github.com/yourusername/authapp-demo"
Repository = "https://github.com/yourusername/authapp-demo.git"

[project.scripts]
authapp-demo = "main:main"
